Update main.py
Added the matrix1() function and addition function. The addition function adds matrixes together and does not seem to be efficient for user input. I found this out online but I feel like there is a better way to do this. Matrix1() will be deleted soon and replaced with other operations. Start() now has user input in our to select the options and is the same as the question for the Temperature converter.